--- pages/404.tsx
@@ -1,11 +1,0 @@
-export default function Page404() {
-    return (
-        <div className="bg-[#1a1a1a] box-border border-0 border-solid border-gray-200 text-white flex h-screen w-screen items-center justify-center m-0 p-0">
-            <div className='flex flex-col items-center justify-center'>
-                <h1 className='text-5xl font-bold'>404</h1>
-                <p className='text-[#8f8f8f]'>Page not found</p>
-
-            </div>
-        </div>
-    )
-}
