+++ src/components/IconCircle.astro
@@ -1,0 +1,17 @@
+<svg
+	class="svg-inline--fa fa-circle"
+	aria-hidden="true"
+	focusable="false"
+	data-prefix="far"
+	data-icon="circle"
+	role="img"
+	xmlns="http://www.w3.org/2000/svg"
+	viewBox="0 0 512 512"
+	data-fa-i2svg=""
+>
+	<path
+		fill="currentColor"
+		d="M464 256A208 208 0 1 0 48 256a208 208 0 1 0 416 0zM0 256a256 256 0 1 1 512 0A256 256 0 1 1 0 256z"
+	>
+	</path>
+</svg>
