+++ src/layouts/Layout.astro
@@ -1,0 +1,49 @@
+---
+// import '@fontsource/inter';
+---
+
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<meta name="viewport" content="width=device-width" />
+		<link rel="icon" type="image/svg+xml" href="/fire_flat.svg" />
+		<meta name="generator" content={Astro.generator} />
+
+		<title>RiskyMH</title>
+
+		{/* SEO */}
+		<meta name="description" content="Just a random person on the internet" />
+		<meta name="keywords" content="RiskyMH" />
+
+		{/* Coloring mainly for discord */}
+		{/* <meta name="theme-color" media="(prefers-color-scheme: light)" content="#2a2a2a" /> */}
+		<meta
+			name="theme-color"
+			media="(prefers-color-scheme: dark)"
+			content="#2a2a2a"
+		/>
+		<meta name="theme-color" content="#FF6723" />
+
+		{/* Fancier SEO */}
+		<meta property="og:title" content="RiskyMH" />
+		<meta
+			property="og:description"
+			content="Just a random person on the internet"
+		/>
+		<meta property="og:type" content="website" />
+		<meta property="og:url" content="https://riskymh.dev/" />
+	</head>
+	<body class="font-sans">
+		<slot />
+	</body>
+</html>
+
+<style is:global>
+	.svg-inline--fa {
+		display: var(--fa-display, inline-block);
+		height: 1em;
+		overflow: visible;
+		vertical-align: -0.125em;
+	}
+</style>
