+++ src/pages/404.astro
@@ -1,0 +1,14 @@
+---
+import Layout from "../layouts/Layout.astro";
+---
+
+<Layout>
+	<div
+		class="bg-[#1a1a1a] box-border border-0 border-solid border-gray-200 text-white flex h-screen w-screen items-center justify-center m-0 p-0"
+	>
+		<div class="flex flex-col items-center justify-center">
+			<h1 class="text-5xl font-bold">404</h1>
+			<p class="text-[#8f8f8f]">Page not found</p>
+		</div>
+	</div>
+</Layout>
