--- styles/globals.css
@@ -1,10 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-.svg-inline--fa {
-  display: var(--fa-display, inline-block);
-  height: 1em;
-  overflow: visible;
-  vertical-align: -.125em;
-}
